+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+# Changelog
+
 ## 0.0.1
 
 - Initial release
-- Added StoryItem.inline feature
-- Basic story view support
+- Added smooth squircle borders and radius support
+- Added shape decoration support for Flutter widgets